+++ api/src/services/og/ogStatsService.ts
@@ -1,6 +1,6 @@
 import { ServiceFactory } from "../../factories/serviceFactory";
 import { INetwork } from "../../models/db/INetwork";
-import { BaseStatsService } from "../baseStatsService";
+import { BaseStatsService } from "../chrysalis/baseStatsService";
 import { OgItemsService } from "./ogItemsService";
 
 /**
